decouple Move 0: credence-skin artifact, wire client, package demotion
Complete the contract: the versioned engine artifact, the one thin client
external apps depend on, and the demotion of the in-process embedding surface.

- Dockerfile.skin -> ghcr.io/gfrmin/credence-skin: the pinnable engine artifact
  (engine + skin only), stdio JSON-RPC wire, precompiled depot.
- Extract credence-skin-client (apps/skin/clients/python): the canonical wire
  client, zero engine code (no juliacall) — cannot host a brain. Adds a `command`
  seam for launching a pinned image (docker run -i) and a $CREDENCE_SKIN_SERVER
  fallback. apps/skin/client.py is now a thin re-export defaulting to the adjacent
  server.jl for in-repo dev. Added to the uv workspace.
- Demote credence + credence-agents to engine-repo-internal (Private :: Do Not
  Upload): in-process embedding is a brain, forbidden as a public library. They
  stay importable for the co-released credence-proxy.
- Write the consumption boundary + co-released-image exception into SPEC §6.9 and
  CLAUDE.md.
- CI: publish-skin + skin-smoke-build jobs (stdio initialize round-trip), and a
  protocol-version invariant step (protocol.md header == PROTOCOL_VERSION).

Verified: 23 credence_router domain tests + skin contract/backward-compat tests
pass through the extracted client; credence-lint check apps/ clean (226 files);
corpus self-test passes.
