+++ SPEC.md
@@ -336,6 +336,12 @@ A baby doesn't know what its hands can do. It learns through motor babbling — 
 
 Higher-level skills emerge not as memorised action sequences but as closed-loop policies — programs that branch on processing state and select the appropriate primitive at each step. "Triage urgent email" is a conditional policy that labels, moves, and notifies based on what's already been done, adapting to interruptions. This follows Sutton et al.'s options framework (1999): options are closed-loop policies, not open-loop sequences. Grammar evolution discovers these policies and compresses them to nonterminals, creating reusable skills. See §7.4.
 
+### 6.9 The consumption boundary: the wire is the only public surface
+
+The brain/body split has a packaging consequence. The engine (the brain machinery: DSL core + the axiom-constrained functions) is consumed by external applications across exactly one sanctioned surface: the **skin wire** (the JSON-RPC protocol of `apps/skin/`, versioned independently of the engine — see `apps/skin/protocol.md`). Across that wire, Measures stay server-side as opaque state IDs; a consumer never holds a measure, so it physically cannot read or do arithmetic on probability weights. The wire boundary therefore enforces Invariant 1's spatial face *by construction* for any application on the far side. External applications declare their domain as **data** — inline BDSL spaces/kernels/priors (`dsl_sources`) — and call the canalised primitives; they carry no probabilistic Julia of their own. The official thin client is `credence-skin-client` (pure JSON-RPC plumbing, zero engine code); the versioned engine artifact is the `credence-skin` image. This is the decoupling commitment: *credence is the engine; an application gets data and a thin body, never a brain, and the engine does not ship the means for an external application to host one.*
+
+**The co-released-image exception.** In-process embedding — loading the engine into the application's own process via `juliacall` (the `credence` / `credence-agents` packages) — *is* hosting a brain, and is forbidden as a published library that external applications import. The two are demoted to engine-repo-internal (`Private :: Do Not Upload`). The single exception: code **co-released inside an engine-repo image** may embed in-process, because there the brain is built and versioned together with the engine, with no application-side embedding — `credence-proxy` (the `credence_router` gateway) and the `credence-pi-daemon` are the standing instances. They are product images shipped from this repo, not libraries others depend on. See `docs/decouple/master-plan.md`.
+
 ---
 
 ## 7. Programs as DSL Expressions
