+++ apps/skin/clients/python/pyproject.toml
@@ -1,0 +1,34 @@
+[build-system]
+requires = ["setuptools>=68.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "credence-skin-client"
+version = "0.1.0"
+description = "Thin JSON-RPC wire client for the Credence skin. Talks the protocol; ships no engine code and cannot host a brain."
+requires-python = ">=3.11"
+readme = "README.md"
+license = "AGPL-3.0-only"
+classifiers = [
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
+    "Topic :: Scientific/Engineering :: Artificial Intelligence",
+]
+# Zero runtime dependencies — pure stdlib (json, subprocess, select). The absence
+# of juliacall is the "cannot host a brain" guarantee made concrete.
+dependencies = []
+
+[project.optional-dependencies]
+dev = [
+    "pytest>=7.4",
+    "ruff",
+]
+
+[project.urls]
+Homepage = "https://github.com/gfrmin/credence"
+Repository = "https://github.com/gfrmin/credence"
+
+[tool.setuptools.packages.find]
+where = ["."]
+include = ["credence_skin_client*"]
